+++ examples/tv-casting-app/tv-casting-common/core/CastingPlayerDiscovery.h
@@ -101,7 +101,7 @@ public:
      * with CastingPlayers whose deviceType matches filterBydeviceType
      * @return CHIP_ERROR - CHIP_NO_ERROR if discovery for CastingPlayers started successfully, specific error code otherwise.
      */
-    CHIP_ERROR StartDiscovery(uint64_t filterBydeviceType = 0);
+    CHIP_ERROR StartDiscovery(uint32_t filterBydeviceType = 0);
 
     /**
      * @brief Stop the discovery for CastingPlayers
